requirements.txt
# Hey Jarvis - Voice Assistant Dependencies
# ==========================================

# Core Voice Assistant
openwakeword>=0.5.0
vosk>=0.3.45
edge-tts>=6.1.0
pyttsx3>=2.90

# Audio Processing
sounddevice>=0.4.6
soundfile>=0.12.1
numpy>=1.24.0
webrtcvad>=2.0.10
pygame>=2.5.0

# LLM Integration
openai>=1.0.0

# Configuration Management
python-dotenv>=1.0.0

# Build & Deployment
pyinstaller>=6.0.0

# Data Processing
scipy>=1.11.0
librosa>=0.10.0

# Utilities
requests>=2.31.0
